+++ index.js
@@ -1,8 +1,12 @@
 const { Celsius } = require('./lib/core')
 const { AUTH_METHODS, ENVIRONMENT } = require('./lib/consts')
+const { ERROR_MESSAGES, CelsiusSDKError, ValidationError } = require('./lib/errors')
 
 module.exports = {
   Celsius,
   AUTH_METHODS,
-  ENVIRONMENT
+  ENVIRONMENT,
+  ERROR_MESSAGES,
+  CelsiusSDKError,
+  ValidationError
 }